+++ src/hapi_schema/db_funding.py
@@ -15,6 +15,7 @@ from sqlalchemy.sql.expression import literal
 
 from hapi_schema.db_location import DBLocation
 from hapi_schema.db_resource import DBResource
+from hapi_schema.utils import endpoint_constants
 from hapi_schema.utils.base import Base
 from hapi_schema.utils.constraints import reference_period_constraint
 from hapi_schema.utils.view_params import ViewParams
@@ -99,8 +100,8 @@ view_params_funding = ViewParams(
 # Results format: category, subcategory, location_name, location_code, admin1_name, admin1_code, admin2_name, admin2_code, hapi_updated_date
 availability_stmt_funding = (
     select(
-        literal("coordination-context").label("category"),
-        literal("funding").label("subcategory"),
+        literal(endpoint_constants.FUNDING_CAT).label("category"),
+        literal(endpoint_constants.FUNDING_SUBCAT).label("subcategory"),
         DBLocation.name.label("location_name"),
         DBLocation.code.label("location_code"),
         literal("").label("admin1_name"),