+++ src/hapi_schema/db_humanitarian_needs.py
@@ -16,6 +16,7 @@ from hapi_schema.db_admin2 import DBAdmin2
 from hapi_schema.db_location import DBLocation
 from hapi_schema.db_resource import DBResource
 from hapi_schema.db_sector import DBSector
+from hapi_schema.utils import endpoint_constants
 from hapi_schema.utils.base import Base
 from hapi_schema.utils.constraints import (
     population_constraint,
@@ -116,8 +117,10 @@ view_params_humanitarian_needs = ViewParams(
 # Results format: category, subcategory, location_name, location_code, admin1_name, admin1_code, admin2_name, admin2_code, hapi_updated_date
 availability_stmt_humanitarian_needs = (
     select(
-        literal("affected-people").label("category"),
-        literal("humanitarian-needs").label("subcategory"),
+        literal(endpoint_constants.HUMANITARIAN_NEEDS_CAT).label("category"),
+        literal(endpoint_constants.HUMANITARIAN_NEEDS_SUBCAT).label(
+            "subcategory"
+        ),
         DBLocation.name.label("location_name"),
         DBLocation.code.label("location_code"),
         DBAdmin1.name.label("admin1_name"),