+++ src/hapi_schema/utils/endpoint_constants.py
@@ -1,0 +1,35 @@
+"""
+The constants below are used to create the category and subcategory names.
+Be careful when changing them as these are used to create the endpoints.
+"""
+
+# categories
+AFFECTED_PEOPLE = "affected-people"
+COORDINATION_CONTEXT = "coordination-context"
+FOOD_SECURITY = "food-security-nutrition-poverty"
+GEOGRAPHY_INFRASTRUCTURE = "geography-infrastructure"
+
+CONFLICT_EVENT_CAT = COORDINATION_CONTEXT
+CONFLICT_EVENT_SUBCAT = "conflict-events"
+FOOD_PRICE_CAT = FOOD_SECURITY
+FOOD_PRICE_SUBCAT = "food-prices-market-monitor"
+FOOD_SECURITY_CAT = FOOD_SECURITY
+FOOD_SECURITY_SUBCAT = "food-security"
+FUNDING_CAT = COORDINATION_CONTEXT
+FUNDING_SUBCAT = "funding"
+HUMANITARIAN_NEEDS_CAT = AFFECTED_PEOPLE
+HUMANITARIAN_NEEDS_SUBCAT = "humanitarian-needs"
+IDPS_CAT = AFFECTED_PEOPLE
+IDPS_SUBCAT = "idps"
+NATIONAL_RISK_CAT = COORDINATION_CONTEXT
+NATIONA_RISK_SUBCAT = "national-risk"
+OPERATIONAL_PRESENCE_CAT = COORDINATION_CONTEXT
+OPERATIONAL_PRESENCE_SUBCAT = "operational-presence"
+POPULATION_CAT = GEOGRAPHY_INFRASTRUCTURE
+POPULATION_SUBCAT = "baseline-population"
+POVERTY_RATE_CAT = FOOD_SECURITY
+POVERTY_RATE_SUBCAT = "poverty-rate"
+REFUGEES_CAT = AFFECTED_PEOPLE
+REFUGEES_SUBCAT = "refugees-persons-of-concern"
+RETURNEES_CAT = AFFECTED_PEOPLE
+RETURNEES_SUBCAT = "returnees"