+++ tests/test_conflict_event.py
@@ -35,7 +35,7 @@ def test_conflict_event_availability(run_view_test):
         view=view_availability,
         whereclause=(
             view_availability.c.category == "coordination-context",
-            view_availability.c.subcategory == "conflict-event",
+            view_availability.c.subcategory == "conflict-events",
             view_availability.c.location_code == "FOO",
             view_availability.c.admin1_name == "Province 01",
             view_availability.c.admin2_name == "District A",