+++ tests/test_returnees.py
@@ -30,8 +30,7 @@ def test_returnees_availability(run_view_test):
         whereclause=(
             view_availability.c.category == "affected-people",
             view_availability.c.subcategory == "returnees",
-            view_availability.c.location_code
-            == "BAR",  # we use the country of asylum, not origin
+            view_availability.c.location_code == "BAR",
             view_availability.c.admin1_name == "",
             view_availability.c.admin2_name == "",
             view_availability.c.hapi_updated_date == datetime(2023, 8, 1),